refactor(Server): housekeeping and organization
