+++ server/route/index.ts
@@ -1,3 +1,1 @@
-export * from './article';
 export * from './router';
-export * from './category';
