+++ server/route/util.ts
@@ -1,0 +1,1 @@
+export const STD_PREFIX = 'http://';
