+++ server/template.env
@@ -1,19 +1,13 @@
-## REST API ##
-
+# Express
 EXPRESS_PORT=
 
-## Database ##
-
-# Offsite deployment
-#DB_CONNECTION_STRING =
-
-## Local Container Deployment
+# Database
 DB_HOST= 
 DB_PORT= 
 DB_NAME= 
 DB_USER=
 DB_PASSWORD= 
 
-# search engine
+# Search engine
 SEARCH_ENGINE_PORT=
 SEARCH_ENGINE_HOST=
